db: fix SetOptions batch range{del,key} refresh logic
The commit a015e5a0c introduced an optimization to avoid refreshing batch
rangedel and rangekey iterators when no rangedel or rangekey mutations were
applied.

This optimization had a bug: It compared unlike numbers to determine whether to
refresh. The batch's `count{RangeDels,RangeKeys}` fields record the number of
rangedels or range keys added to the batch and are monotonically increasing.
The count of spans in the `keyspan.Iter` on the other hand was the count of
fragments. This bug could result in erroroneously not refreshing the iterator's
view if the previous number of fragments matched the current number of batch
keys.

This commit adapts SetOptions to not avoid calls to initRangeDelIter and
initRangeKeyIter in these cases: These functions already avoid refragmenting
the spans if unnecessary, so there should be little cost to calling them
unconditionally. The logic to reconstruct the point / range key iterator in
SetOptions when a batch goes from zero rangedels/rangekeys to nonzero
rangedels/rangekeys is preserved.

+++ batch.go
@@ -216,11 +216,11 @@ type Batch struct {
 
 	// The count of range deletions in the batch. Updated every time a range
 	// deletion is added.
-	countRangeDels uint64
+	countRangeDels uint32
 
 	// The count of range key sets, unsets and deletes in the batch. Updated
 	// every time a RANGEKEYSET, RANGEKEYUNSET or RANGEKEYDEL key is added.
-	countRangeKeys uint64
+	countRangeKeys uint32
 
 	// A deferredOp struct, stored in the Batch so that a pointer can be returned
 	// from the *Deferred() methods rather than a value.
